Put extract credits function into state module
